src/environments/sc2/requirements.txt
astroid==2.2.5
atari-py==0.1.15
autopep8==1.4.4
future==0.17.1
gym==0.12.5
gym-retro==0.7.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pycodestyle==2.5.0
pyglet==1.3.2
pylint==2.3.1
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
s2clientprotocol==4.9.1.74456.0
websocket-client==0.56.0